Networking Now Actually Works!
Well... that may be stretching it a bit...

At least you can move a sprite over an internet connection via the
custom packet protocol!

Note: There may be an infinite loop / memory leak in the parsing for an
update packet. This is due to either a bad loop, or the JPacketFieldX
enum is off (again).